# packages/contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: «Инспектор ИИ» — REST API
  version: 0.1.0
  description: |
    Контракт REST API веб-продукта «Инспектор ИИ» (M1). Документ ведёт AG-00 (контракты); владелец каждой
    операции указан в `x-owner`.

    Правила (docs/analysis/90 §3.4):
    - JSON snake_case, перечисления — строки из `packages/contracts/enums.yaml`
      (поле `x-contract-enum` указывает перечисление; совпадение проверяется тестом).
    - Ошибки — `application/problem+json` (RFC 9457) по `packages/contracts/schemas/problem.schema.json`,
      коды — из `packages/contracts/errors.yaml`.
    - Каждый ответ несёт заголовок `X-Request-Id` (переданный клиентом или сгенерированный UUID v7).
    - Запросы проверяются по этому документу; в режимах local/test проверяются и ответы.
    - Веб-продукт импортирует каталоги запусков `inspector-batch` (97 §2.14); решения инспектора
      никогда не попадают в выгрузку для организаторов.
    - Вход по логину и паролю (09 §3.3.1): сессия в cookie `ii_sid` (HttpOnly, SameSite=Strict; `__Host-ii_sid`
      при INSPECTOR_COOKIE_SECURE), для POST/PUT/PATCH/DELETE — заголовок `X-CSRF-Token` из ответа входа.
      `security: []` — операция доступна без входа; `x-permission` — право из `packages/contracts/rbac.yaml`
      (роль → право → область ALL | ASSIGNED | OWN); `x-allow-password-change: true` — операция доступна до
      обязательной смены пароля; `x-audit-action` — действие журнала аудита для изменяющей операции.
...
  x-owner: AG-00
servers:
...
security:
  - sessionCookie: []
paths:
...
  /objects:
    get:
      operationId: listObjects
      tags: [objects]
      summary: Список объектов с количеством файлов по стадиям
      x-owner: AG-08
      x-permission: object.read
      x-error-codes: [VALIDATION_ERROR, DB_UNAVAILABLE, UNAUTHENTICATED, SESSION_EXPIRED, PASSWORD_CHANGE_REQUIRED, FORBIDDEN, SESSION_STORE_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: q
          in: query
          description: Поиск по идентификатору или наименованию объекта (без учёта регистра)
          schema: {type: string, minLength: 1, maxLength: 200}
        - name: include_archived
          in: query
          description: Показывать и архивные объекты («Архивировать»); по умолчанию они скрыты
          schema: {type: boolean}
      responses:
        '200':
          description: Страница списка объектов (сортировка по object_id)
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ObjectList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /objects/{object_id}:
    get:
      operationId: getObject
      tags: [objects]
      summary: Карточка объекта
      x-owner: AG-08
      x-permission: object.read
      x-error-codes: [VALIDATION_ERROR, NOT_FOUND, DB_UNAVAILABLE, UNAUTHENTICATED, SESSION_EXPIRED, PASSWORD_CHANGE_REQUIRED, FORBIDDEN, SESSION_STORE_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/ObjectIdPath'
      responses:
        '200':
          description: Объект
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ObjectDetail'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /objects/{object_id}/files:
    get:
      operationId: listObjectFiles
      tags: [objects]
      summary: Файлы объекта (реестр документов)
      x-owner: AG-08
      x-permission: object.read
      x-error-codes: [VALIDATION_ERROR, NOT_FOUND, DB_UNAVAILABLE, UNAUTHENTICATED, SESSION_EXPIRED, PASSWORD_CHANGE_REQUIRED, FORBIDDEN, SESSION_STORE_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/ObjectIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: stage
          in: query
          description: Фильтр по стадии из манифеста
          schema: {$ref: '#/components/schemas/ManifestStage'}
        - name: local_status
          in: query
          description: Фильтр по наличию файла на диске
          schema: {$ref: '#/components/schemas/LocalFileStatus'}
        - name: q
          in: query
          description: Поиск по file_id, имени файла или относительному пути (без учёта регистра)
          schema: {type: string, minLength: 1, maxLength: 200}
      responses:
        '200':
          description: Страница файлов объекта (сортировка по file_id)
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FileList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
  /objects/{object_id}/protocols:
    get:
      operationId: listObjectProtocols
      tags: [protocols]
      summary: Версии протокола объекта (по импортированным запускам, новые сверху)
      x-owner: AG-08
      x-permission: protocol.read
      x-error-codes: [VALIDATION_ERROR, SESSION_STORE_UNAVAILABLE, FORBIDDEN, PASSWORD_CHANGE_REQUIRED, SESSION_EXPIRED, UNAUTHENTICATED, NOT_FOUND, HIDDEN_TEST_ACCESS_DENIED, CONTRACT_VALIDATION_FAILED, DB_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/ObjectIdPath'
      responses:
        '200':
          description: История версий протокола
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProtocolVersionList'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /objects/{object_id}/protocols/{run_id}:
    get:
      operationId: getObjectProtocol
      tags: [protocols]
      summary: Протокол (контракт protocol.schema.json) и группы находок того же запуска
      x-owner: AG-08
      x-permission: protocol.read
      x-error-codes: [VALIDATION_ERROR, SESSION_STORE_UNAVAILABLE, FORBIDDEN, PASSWORD_CHANGE_REQUIRED, SESSION_EXPIRED, UNAUTHENTICATED, NOT_FOUND, HIDDEN_TEST_ACCESS_DENIED, CONTRACT_VALIDATION_FAILED, DB_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/ObjectIdPath'
        - $ref: '#/components/parameters/RunIdPath'
      responses:
        '200':
          description: Протокол
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ProtocolView'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /objects/{object_id}/protocols/{run_id}/export:
    get:
      operationId: exportObjectProtocol
      tags: [protocols]
      summary: Выгрузка протокола (DOCX и PDF формирует inspector-batch export, AG-04; JSON — канонический протокол)
      x-owner: AG-08
      x-permission: protocol.export
      x-error-codes: [VALIDATION_ERROR, SESSION_STORE_UNAVAILABLE, FORBIDDEN, PASSWORD_CHANGE_REQUIRED, SESSION_EXPIRED, UNAUTHENTICATED, NOT_FOUND, HIDDEN_TEST_ACCESS_DENIED, DB_UNAVAILABLE, INTERNAL_ERROR]
      parameters:
        - $ref: '#/components/parameters/ObjectIdPath'
        - $ref: '#/components/parameters/RunIdPath'
        - name: format
          in: query
          required: true
          description: Формат файла
          schema: {$ref: '#/components/schemas/ProtocolFileFormat'}
      responses:
        '200':
          description: Файл протокола (Content-Disposition attachment, имя UTF-8)
          headers:
            X-Request-Id: {$ref: '#/components/headers/XRequestId'}
            Content-Disposition:
              description: attachment; filename*=UTF-8''Протокол_….docx
              schema: {type: string}
          content:
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema: {type: string, format: binary}
            application/pdf:
              schema: {type: string, format: binary}
            application/json:
              schema: {type: string, format: binary}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/AccessDenied'}
        '404': {$ref: '#/components/responses/NotFound'}
        '500': {$ref: '#/components/responses/InternalError'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: ii_sid
      description: Непрозрачный токен сессии (256 бит); на сервере хранится только его sha256 (Redis). При INSPECTOR_COOKIE_SECURE — `__Host-ii_sid`.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: csrf_token из ответа входа или GET /auth/me; обязателен для POST/PUT/PATCH/DELETE при входе по сессии.
  headers:
...
    XRequestId:
      description: Идентификатор запроса (переданный клиентом или UUID v7); тот же, что в логах и в теле ошибки
      schema: {type: string}
...
  parameters:
...
    ObjectIdPath:
      name: object_id
      in: path
      required: true
      description: Идентификатор объекта (object_id манифеста)
      schema: {type: string, minLength: 1, maxLength: 64}
    Page:
      name: page
      in: query
      description: Номер страницы, с 1
      schema: {type: integer, minimum: 1, maximum: 100000, default: 1}
    PageSize:
      name: page_size
      in: query
      description: Размер страницы (не более 200)
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
    RunIdPath:
      name: run_id
      in: path
      required: true
      description: Внутренний идентификатор импортированного запуска (UUID v7, поле id в /admin/batch-runs)
      schema: {type: string, format: uuid}
...
  responses:
    BadRequest:
      description: Запрос не прошёл проверку схемы (VALIDATION_ERROR, MALFORMED_JSON)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: Ресурс не найден (NOT_FOUND)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    Unprocessable:
      description: Запрос не может быть выполнен (CONTRACT_VALIDATION_FAILED, PASSWORD_POLICY_VIOLATION, FILE_NOT_RENDERABLE)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: Требуется вход или сессия истекла (UNAUTHENTICATED, SESSION_EXPIRED, INVALID_CREDENTIALS, REAUTH_REQUIRED)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    AccessDenied:
      description: Недостаточно прав, неверный токен формы или требуется смена пароля (FORBIDDEN, CSRF_TOKEN_INVALID, PASSWORD_CHANGE_REQUIRED, HIDDEN_TEST_ACCESS_DENIED)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    InternalError:
      description: Внутренняя ошибка (INTERNAL_ERROR, RESPONSE_SCHEMA_VIOLATION)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unavailable:
      description: Сервис временно недоступен (DB_UNAVAILABLE, SESSION_STORE_UNAVAILABLE, PAGE_RENDERER_UNAVAILABLE)
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
...
    ManifestStage:
      type: string
      x-contract-enum: ManifestStage
      enum: [PD, RD, ID, RD_ID_MIXED, UNKNOWN]
    DocStageOrNull:
      type: string
      nullable: true
      x-contract-enum: DocStage
      enum: [PD, RD, ID, null]
...
    ManifestSplitOrNull:
      type: string
      nullable: true
      x-contract-enum: ManifestSplit
      enum: [TRAIN_PUBLIC, TEST_HIDDEN, null]
    LocalFileStatus:
      type: string
      x-contract-enum: LocalFileStatus
      enum: [PRESENT, RECOVERED, MISSING_ON_DISK]
    LoadScenarioOrNull:
      type: string
      nullable: true
      x-contract-enum: LoadScenario
      enum: [FULL, PD_RD_ONLY, PD_ID_ONLY, RD_ID_ONLY, SINGLE_ONLY, PARTIALLY_LOADED, null]
    IndicatorColor:
      type: string
      x-contract-enum: IndicatorColor
      enum: [RED, YELLOW, GREEN, NONE]
...
    ProtocolStatus:
      type: string
      x-contract-enum: ProtocolStatus
      enum: [IN_VERIFICATION, VERIFICATION_COMPLETED, PROTOCOL_FINALIZED, SUPERSEDED]
...
    Sha256:
      type: string
      pattern: '^[0-9a-f]{64}$'
    Timestamp:
      type: string
      format: date-time
    # ── Health ──
...
    StageCounts:
      type: object
      description: Количество файлов по стадии из манифеста (ключи — ManifestStage)
      additionalProperties: false
      required: [PD, RD, ID, RD_ID_MIXED, UNKNOWN]
      properties:
        PD: {type: integer, minimum: 0}
        RD: {type: integer, minimum: 0}
        ID: {type: integer, minimum: 0}
        RD_ID_MIXED: {type: integer, minimum: 0}
        UNKNOWN: {type: integer, minimum: 0}
    RunRef:
      type: object
      nullable: true
      additionalProperties: false
      required: [id, batch_run_id, imported_at]
      properties:
        id: {type: string, format: uuid}
        batch_run_id: {type: string}
        imported_at: {$ref: '#/components/schemas/Timestamp'}
    ObjectSummary:
      type: object
      additionalProperties: false
      required:
        - object_id
        - name
        - split
        - scenario
        - indicator_color
        - files_total
        - files_present
        - files_missing_on_disk
        - files_by_stage
        - last_run
        - updated_at
      properties:
        object_id: {type: string, minLength: 1, maxLength: 64}
        name: {type: string, nullable: true, description: 'Наименование (поле corpus манифеста организаторов)'}
        split: {$ref: '#/components/schemas/ManifestSplitOrNull'}
        scenario: {$ref: '#/components/schemas/LoadScenarioOrNull'}
        indicator_color: {$ref: '#/components/schemas/IndicatorColor'}
        files_total: {type: integer, minimum: 0}
        files_present: {type: integer, minimum: 0, description: 'Файлы со статусом PRESENT или RECOVERED'}
        files_missing_on_disk: {type: integer, minimum: 0}
        files_by_stage: {$ref: '#/components/schemas/StageCounts'}
        last_run: {$ref: '#/components/schemas/RunRef'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
    ObjectDetail:
      type: object
      additionalProperties: false
      description: ObjectSummary и поля карточки объекта (ТЗ §10.3; пока не заполняются импортом)
      required:
        - object_id
        - name
        - split
        - scenario
        - indicator_color
        - files_total
        - files_present
        - files_missing_on_disk
        - files_by_stage
        - last_run
        - updated_at
        - input_manifest_hash
        - address
        - customer
        - contractor
        - permit_number
        - created_at
      properties:
        object_id: {type: string, minLength: 1, maxLength: 64}
        name: {type: string, nullable: true, description: 'Наименование (поле corpus манифеста организаторов)'}
        split: {$ref: '#/components/schemas/ManifestSplitOrNull'}
        scenario: {$ref: '#/components/schemas/LoadScenarioOrNull'}
        indicator_color: {$ref: '#/components/schemas/IndicatorColor'}
        files_total: {type: integer, minimum: 0}
        files_present: {type: integer, minimum: 0, description: 'Файлы со статусом PRESENT или RECOVERED'}
        files_missing_on_disk: {type: integer, minimum: 0}
        files_by_stage: {$ref: '#/components/schemas/StageCounts'}
        last_run: {$ref: '#/components/schemas/RunRef'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
        input_manifest_hash: {type: string, nullable: true, pattern: '^[0-9a-f]{64}$'}
        address: {type: string, nullable: true}
        customer: {type: string, nullable: true}
        contractor: {type: string, nullable: true}
        permit_number: {type: string, nullable: true}
        created_at: {$ref: '#/components/schemas/Timestamp'}
    ObjectList:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/ObjectSummary'}
        total: {type: integer, minimum: 0}
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 200}
    FileItem:
      type: object
      additionalProperties: false
      required:
        - file_id
        - object_id
        - file_name
        - relative_path
        - extension
        - size_bytes
        - pdf_pages
        - sha256
        - manifest_stage
        - stage_resolved
        - manifest_section
        - dataset_role
        - split
        - annotation_status
        - duplicate_group
        - local_status
        - sha256_verified
        - last_run_id
        - updated_at
      properties:
        file_id: {type: string, minLength: 1, maxLength: 512}
        object_id: {type: string}
        file_name: {type: string}
        relative_path: {type: string, nullable: true, description: 'Путь относительно 01_ДОКУМЕНТАЦИЯ (manifest relative_path)'}
        extension: {type: string, nullable: true}
        size_bytes: {type: integer, minimum: 0, nullable: true}
        pdf_pages: {type: integer, minimum: 0, nullable: true}
        sha256: {$ref: '#/components/schemas/Sha256'}
        manifest_stage: {$ref: '#/components/schemas/ManifestStage'}
        stage_resolved: {$ref: '#/components/schemas/DocStageOrNull'}
        manifest_section: {type: string, nullable: true, description: 'ManifestSection (открытый словарь организаторов)'}
        dataset_role: {type: string, nullable: true, description: 'DatasetRole (открытый словарь организаторов)'}
        split: {$ref: '#/components/schemas/ManifestSplitOrNull'}
        annotation_status: {type: string, nullable: true, description: 'AnnotationStatus (открытый словарь организаторов)'}
        duplicate_group: {type: string, nullable: true}
        local_status: {$ref: '#/components/schemas/LocalFileStatus'}
        sha256_verified: {type: boolean, nullable: true}
        last_run_id: {type: string, format: uuid, nullable: true}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
    FileList:
      type: object
      additionalProperties: false
      required: [items, total, page, page_size]
      properties:
        items:
          type: array
          items: {$ref: '#/components/schemas/FileItem'}
        total: {type: integer, minimum: 0}
        page: {type: integer, minimum: 1}
        page_size: {type: integer, minimum: 1, maximum: 200}
    # ── Batch runs ──
...
    ProblemItem:
      type: object
      required: [code, title, detail]
      properties:
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]*$'}
        title: {type: string}
        detail: {type: string}
        details: {type: object}
        pointer: {type: string}
        file_name: {type: string}
        row: {type: integer}
        field: {type: string}
        params: {type: object}
    Problem:
      type: object
      required: [type, title, status, detail, instance, code, request_id, timestamp, retryable]
      properties:
        type: {type: string, pattern: '^/problems/[a-z0-9-]+$'}
        title: {type: string, minLength: 1}
        status: {type: integer, minimum: 400, maximum: 599}
        detail: {type: string, minLength: 1}
        instance: {type: string}
        code: {type: string, pattern: '^[A-Z][A-Z0-9_]*$'}
        request_id: {type: string, minLength: 1}
        timestamp: {$ref: '#/components/schemas/Timestamp'}
        retryable: {type: boolean}
        retry_after_s: {type: integer, minimum: 0}
        details: {type: object}
        errors:
          type: array
          items: {$ref: '#/components/schemas/ProblemItem'}
        warnings:
          type: array
          items: {$ref: '#/components/schemas/ProblemItem'}
...
    LoadScenario:
      type: string
      x-contract-enum: LoadScenario
      enum: [FULL, PD_RD_ONLY, PD_ID_ONLY, RD_ID_ONLY, SINGLE_ONLY, PARTIALLY_LOADED]
...
    ProtocolFileFormat:
      type: string
      description: Подмножество ExportFormat, которое выгружает inspector-batch export
      enum: [json, docx, pdf]
...
    ProtocolCounts:
      type: object
      additionalProperties: false
      required: [critical, substantial, suspicions, not_checked_no_id, cards]
      properties:
        critical: {type: integer, minimum: 0}
        substantial: {type: integer, minimum: 0}
        suspicions: {type: integer, minimum: 0}
        not_checked_no_id: {type: integer, minimum: 0}
        cards: {type: integer, minimum: 0}
    ProtocolVersion:
      type: object
      additionalProperties: false
      required:
        - run_id
        - batch_run_id
        - object_id
        - protocol_no
        - version
        - web_version
        - is_latest
        - generated_at
        - status
        - is_final
        - scenario
        - status_line
        - counts
        - formats
        - pipeline_version
        - matrix_version
        - input_manifest_hash
        - content_sha256
      properties:
        run_id: {type: string, format: uuid, description: 'id импортированного запуска'}
        batch_run_id: {type: string}
        object_id: {type: string}
        protocol_no: {type: string}
        version: {type: integer, minimum: 1, description: 'Версия из протокола (поле version)'}
        web_version: {type: integer, minimum: 1, description: 'Порядковый номер в истории объекта (1 — самый ранний)'}
        is_latest: {type: boolean}
        generated_at: {$ref: '#/components/schemas/Timestamp'}
        status: {$ref: '#/components/schemas/ProtocolStatus'}
        is_final: {type: boolean}
        scenario: {$ref: '#/components/schemas/LoadScenario'}
        status_line: {type: string}
        counts: {$ref: '#/components/schemas/ProtocolCounts'}
        formats:
          type: array
          description: Доступные файлы выгрузки
          items: {$ref: '#/components/schemas/ProtocolFileFormat'}
        pipeline_version: {type: string}
        matrix_version: {type: string, nullable: true}
        input_manifest_hash: {$ref: '#/components/schemas/Sha256'}
        content_sha256: {type: string, nullable: true, pattern: '^[0-9a-f]{64}$'}
    ProtocolVersionList:
      type: object
      additionalProperties: false
      required: [object_id, items]
      properties:
        object_id: {type: string}
        items:
          type: array
          items: {$ref: '#/components/schemas/ProtocolVersion'}
    ContractProtocol:
      type: object
      description: Канонический протокол — packages/contracts/schemas/protocol.schema.json (проверяется API перед выдачей)
      x-contract-schema: protocol
      required: [schema_version, protocol_no, object, version, is_final, status, generated_at, scenario, versions, input_manifest_hash, header, appendix2, tz92_tables, evidence_cards, input_registry]
      additionalProperties: true
    ContractFindingGroup:
      type: object
      description: Группа находок — packages/contracts/schemas/finding_group.schema.json (проверяется API перед выдачей)
      x-contract-schema: finding_group
      required: [finding_group_id, object_id, matrix_scope, parameter_code, comparison_result, location_type, locations, violation_label, protocol_status, criticality, anchor_evidence, finding_ids]
      additionalProperties: true
    ProtocolView:
      type: object
      additionalProperties: false
      required: [version, protocol, finding_groups, warnings]
      properties:
        version: {$ref: '#/components/schemas/ProtocolVersion'}
        protocol: {$ref: '#/components/schemas/ContractProtocol'}
        finding_groups:
          type: array
          items: {$ref: '#/components/schemas/ContractFindingGroup'}
        warnings:
          type: array
          items: {$ref: '#/components/schemas/ProblemItem'}
    # ── Page annotations (layout artifacts, AG-02B) ──
...